Add: グループ成績に大会フィルターを追加し、統計ロジックをconcernに抽出
