+++ app/controllers/api/v1/groups_controller.rb
@@ -1,6 +1,7 @@
 module Api
   module V1
     class GroupsController < ApplicationController
+      include GroupStatsBuilder
       before_action :authenticate_api_v1_user!, only: %i[show create update destroy invite_link]
 
       def index
@@ -114,28 +115,6 @@ module Api
       end
 
       private
-
-      def build_group_stats(accepted_users)
-        year = params[:year]
-        match_type = params[:match_type]
-
-        batting_averages = accepted_users.map { |u| BattingAverage.filtered_aggregate_for_user(u.id, year:, match_type:) }
-        batting_stats = accepted_users.map { |u| BattingAverage.stats_for_user(u.id, year:, match_type:) }
-        pitching_aggregate = accepted_users.map { |u| PitchingResult.filtered_pitching_aggregate_for_user(u.id, year:, match_type:) }
-        pitching_stats = accepted_users.map { |u| PitchingResult.pitching_stats_for_user(u.id, year:, match_type:) }
-        available_years = fetch_available_years(accepted_users)
-
-        { batting_averages:, batting_stats:, pitching_aggregate:, pitching_stats:, available_years: }
-      end
-
-      def fetch_available_years(accepted_users)
-        user_ids = accepted_users.map(&:id)
-        MatchResult.joins(:game_result)
-                   .where(game_results: { user_id: user_ids })
-                   .select('EXTRACT(YEAR FROM date_and_time) AS year')
-                   .distinct.order(Arel.sql('year DESC'))
-                   .map { |r| r.year.to_i }
-      end
 
       def group_params
         params.require(:group).permit(:name, :icon)
